refactor(FileOperations, Files): remove server package dependency by integrating utility functions directly and updating file handling logic
