+++ beam/server/handlers/types.go
@@ -1,4 +1,4 @@
-package server
+package handlers
 
 import (
 	"fmt"
